# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 自炊支援・食費節約アプリ API
  description: |
    一人暮らしまたは2〜4人家族の世帯を対象とした自炊支援・食費節約Webアプリケーションのバックエンド API仕様書。
    
    ## 主要機能
    - ユーザー管理（登録、ログイン、プロフィール管理）
    - レシピ管理（作成、検索、編集、削除）
    - レビュー機能（星評価、コメント、通報）
    - スケジュール管理（料理予定と実績）
    - 買い物リスト管理
    - サボり防止アラート
    - 管理者機能
    
    ## 認証
    - AWS Cognito認証を使用
    - アクセストークンの有効期限: 1時間
    - リフレッシュトークンの有効期限: 90日
    - ユーザーIDはJWTトークンのsubクレームから取得（SecurityContext）
    
    ## 多言語対応
    - 日本語（ja）、韓国語（ko）
    - Accept-Languageヘッダーでエラーメッセージの言語を指定
    
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Private
servers:
  - url: http://localhost:8080
    description: ローカル開発環境
  - url: https://api-dev.example.com
    description: 開発環境
  - url: https://api.example.com
    description: 本番環境
...
paths:
...
  /api/recipes:
    get:
      tags: [Recipes]
      summary: レシピ検索
      description: レシピを検索します。
      operationId: searchRecipes
      security: []
      parameters:
        - name: keyword
          in: query
          description: 検索キーワード
          schema:
            type: string
        - name: authorId
          in: query
          description: 作成者ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 検索成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipeResponse'
    post:
      tags: [Recipes]
      summary: レシピ作成
      description: 新しいレシピを作成します。ユーザーIDはJWTトークンから取得されます。
      operationId: createRecipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeRequest'
      responses:
        '201':
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/recipes/{id}:
    get:
      tags: [Recipes]
      summary: レシピ詳細取得
      description: レシピの詳細情報を取得します。
      operationId: getRecipe
      security: []
      parameters:
        - $ref: '#/components/parameters/RecipeIdPath'
      responses:
        '200':
          description: 取得成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Recipes]
      summary: レシピ更新
      description: レシピを更新します。ユーザーIDはJWTトークンから取得されます。
      operationId: updateRecipe
      parameters:
        - $ref: '#/components/parameters/RecipeIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Recipes]
      summary: レシピ削除
      description: レシピを論理削除します。ユーザーIDはJWTトークンから取得されます。
      operationId: deleteRecipe
      parameters:
        - $ref: '#/components/parameters/RecipeIdPath'
      responses:
        '204':
          description: 削除成功
        '404':
          $ref: '#/components/responses/NotFound'
  /api/recipes/{id}/image:
    post:
      tags: [Recipes]
      summary: レシピ画像アップロード
      description: レシピ画像をアップロードします（JPEG/PNG、最大5MB）。ユーザーIDはJWTトークンから取得されます。
      operationId: uploadRecipeImage
      parameters:
        - $ref: '#/components/parameters/RecipeIdPath'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: アップロード成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/recipes/{id}/steps/{stepIndex}/image:
    post:
      tags: [Recipes]
      summary: 手順画像アップロード
      description: 手順画像をアップロードします（JPEG/PNG、最大5MB）。ユーザーIDはJWTトークンから取得されます。
      operationId: uploadStepImage
      parameters:
        - $ref: '#/components/parameters/RecipeIdPath'
        - name: stepIndex
          in: path
          required: true
          description: 手順インデックス（0始まり）
          schema:
            type: integer
            minimum: 0
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: アップロード成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/recipes/search/by-ingredients:
    post:
      tags: [Recipes]
      summary: 食材でレシピ検索
      description: |
        指定した食材をすべて含むレシピを検索します（AND条件）。
        RecipeIngredientsテーブル（逆引きインデックス）を使用して効率的に検索します。
      operationId: searchRecipesByIngredients
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchByIngredientsRequest'
      responses:
        '200':
          description: 検索成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeSearchByIngredientResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ========================================
  # Reviews API
  # ========================================
...
  /api/admin/recipes/{recipeId}:
    delete:
      tags: [Admin]
      summary: レシピ削除（管理者用）
      description: レシピを論理削除します（参照は保持）。
      operationId: deleteRecipeByAdmin
      parameters:
        - $ref: '#/components/parameters/AdminRecipeIdPath'
      responses:
        '204':
          description: 削除成功
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
...
  parameters:
...
    RecipeIdPath:
      name: id
      in: path
      required: true
      description: レシピID
      schema:
        type: string
        format: uuid
...
    AdminRecipeIdPath:
      name: recipeId
      in: path
      required: true
      description: レシピID
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: バリデーションエラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: 権限エラー
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: リソースが見つからない
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CreateRecipeRequest:
      type: object
      required: [title, ingredients, steps, cookingTime]
      properties:
        title:
          type: string
          maxLength: 100
          example: 簡単カレーライス
        ingredients:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Ingredient'
        steps:
          type: array
          minItems: 1
          items:
            type: string
          example: ["野菜を切る", "炒める", "煮込む"]
        cookingTime:
          type: integer
          minimum: 0
          description: 調理時間（分）
          example: 30
    Ingredient:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
          example: 玉ねぎ
        quantity:
          type: number
          minimum: 0.01
          maximum: 9999
          description: 数量（任意）
          example: 2
        unit:
          type: string
          maxLength: 50
          description: 単位（自由入力、任意）
          example: 個
        note:
          type: string
          maxLength: 200
          example: 中サイズ
        optional:
          type: boolean
          default: false
    RecipeResponse:
      type: object
      properties:
        recipeId:
          type: string
          format: uuid
        title:
          type: string
        authorId:
          type: string
          format: uuid
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
        steps:
          type: array
          items:
            $ref: '#/components/schemas/Step'
        cookingTime:
          type: integer
          description: 調理時間（分）
        imageUrl:
          type: string
          format: uri
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Step:
      type: object
      required: [description]
      properties:
        description:
          type: string
          description: 手順の説明
          example: 野菜を一口大に切る
        imageUrl:
          type: string
          format: uri
          description: 手順の画像URL（任意）
        videoUrl:
          type: string
          format: uri
          description: 手順の動画URL（任意、YouTube等に対応）
          example: https://www.youtube.com/watch?v=...
    SearchByIngredientsRequest:
      type: object
      required: [ingredients]
      properties:
        ingredients:
          type: array
          minItems: 1
          items:
            type: string
          description: 検索する食材名のリスト
          example: ["玉ねぎ", "じゃがいも", "にんじん"]
    RecipeSearchByIngredientResponse:
      type: object
      properties:
        recipes:
          type: array
          items:
            $ref: '#/components/schemas/RecipeSummary'
        totalCount:
          type: integer
          description: 検索結果の総数
          example: 5
        searchedIngredients:
          type: array
          items:
            type: string
          description: 検索に使用した食材リスト
          example: ["玉ねぎ", "じゃがいも"]
    RecipeSummary:
      type: object
      properties:
        recipeId:
          type: string
          format: uuid
        title:
          type: string
          example: 簡単カレーライス
        imageUrl:
          type: string
          format: uri
          nullable: true

    # ----------------------------------------
    # Review Schemas
    # ----------------------------------------
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: バリデーションエラー
        path:
          type: string
          example: /api/users/register

  # ========================================
  # Security Schemes
  # ========================================
  securitySchemes:
    CognitoAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        AWS Cognito JWTトークン認証。
        ログインエンドポイント（/api/users/login）から取得したアクセストークンを使用します。
        ユーザーIDはJWTトークンのsubクレームから自動的に取得されます（SecurityContext）。
        
        例: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`

# ========================================
# Global Security (認証不要のエンドポイントは個別にsecurity: []を指定)
# ========================================
security:
  - CognitoAuth: []